+++ packages/core/admin/admin/src/content-manager/pages/EditView/index.js
@@ -27,30 +27,16 @@ import { getTrad } from '../../utils';
 import DraftAndPublishBadge from './DraftAndPublishBadge';
 import Informations from './Informations';
 import Header from './Header';
-import { createAttributesLayout, getFieldsActionMatchingPermissions } from './utils';
+import {
+  createAttributesLayout,
+  getFieldsActionMatchingPermissions,
+  getCustomFieldUidsFromLayout,
+} from './utils';
 import DeleteLink from './DeleteLink';
 import useLazyComponents from '../../hooks/useLazyComponents';
 
 const cmPermissions = permissions.contentManager;
 const ctbPermissions = [{ action: 'plugin::content-type-builder.read', subject: null }];
-
-const getCustomFieldUidsFromLayout = (layout) => {
-  // Get all the fields on the content-type and its components
-  const allFields = [
-    ...layout.contentType.layouts.edit,
-    ...Object.values(layout.components).flatMap((component) => component.layouts.edit),
-  ].flat();
-  // Filter that down to custom fields and map the uids
-  const customFieldUids = allFields
-    .filter((field) => field.fieldSchema.customField)
-    .map((customField) => customField.fieldSchema.customField);
-  // Make sure the list is unique
-  const uniqueCustomFieldUids = customFieldUids.filter(
-    (uid, index) => customFieldUids.indexOf(uid) === index
-  );
-
-  return uniqueCustomFieldUids;
-};
 
 /* eslint-disable  react/no-array-index-key */
 const EditView = ({
