+++ packages/core/admin/admin/src/content-manager/pages/EditView/utils/index.js
@@ -1,3 +1,4 @@
 // eslint-disable-next-line import/prefer-default-export
 export { default as createAttributesLayout } from './createAttributesLayout';
 export { default as getFieldsActionMatchingPermissions } from './getFieldsActionMatchingPermissions';
+export { default as getCustomFieldUidsFromLayout } from './getCustomFieldUidsFromLayout';
